+++ src/OpenTelemetry.Instrumentation.AspNetCore/CHANGELOG.md
@@ -2,13 +2,25 @@
 
 ## Unreleased
 
+* The `http.server.request.duration` histogram (measured in seconds) produced by
+  the metrics instrumentation in this package now uses the [Advice API](https://github.com/open-telemetry/opentelemetry-dotnet/blob/core-1.10.0/docs/metrics/customizing-the-sdk/README.md#explicit-bucket-histogram-aggregation)
+  to set default explicit buckets following the [OpenTelemetry Specification](https://github.com/open-telemetry/semantic-conventions/blob/v1.29.0/docs/http/http-metrics.md).
+  ([#2430](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2430))
+
+## 1.10.1
+
+Released 2024-Dec-10
+
+* Metric support for the .NET Standard target was removed by mistake in 1.10.0.
+  This functionality has been restored.
+  ([#2403](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2403))
+
 ## 1.10.0
 
 Released 2024-Dec-09
 
 * Drop support for .NET 6 as this target is no longer supported.
-  ([#2138](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2138),
-  ([#2360](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2360))
+  ([#2138](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2138))
 
 * Updated OpenTelemetry core component version(s) to `1.10.0`.
   ([#2317](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2317))